完成Linux移植

+++ test/utest/file_test.cpp
@@ -21,46 +21,46 @@ ASL_UTEST_CASE(file, file_test) {
 	ASL_UTEST_ASSERT_TRUE(file.Open(szFilename, File::OF_Out));
 	ASL_UTEST_ASSERT_EQ(file.Write(acBuffer1, nBufferSize), nBufferSize);
 	ASL_UTEST_ASSERT_EQ((int)file.GetFileSize(), nBufferSize);
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)nBufferSize);
 	file.Close();
 
 	ASL_UTEST_ASSERT_TRUE(file.Open(szFilename, File::OF_In));
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), 0);
-	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)0);
+	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), (size_t)nBufferSize);
 	ASL_UTEST_ASSERT_EQ(file.Read(acBuffer2, nBufferSize), nBufferSize);
 	ASL_UTEST_ASSERT_EQ(memcmp(acBuffer1, acBuffer2, nBufferSize), 0);
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)nBufferSize);
 	ASL_UTEST_ASSERT_EQ(file.Read(acBuffer2, nBufferSize), 0);
 	file.Close();
 
 	ASL_UTEST_ASSERT_TRUE(file.Open(szFilename, File::OF_Out | File::OF_App));
-	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), (size_t)nBufferSize);
 	//ASL_UTEST_ASSERT_EQ(file.GetPos(), nBufferSize);
 	ASL_UTEST_ASSERT_EQ(file.Write(acBuffer1, nBufferSize), nBufferSize);
-	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), 2 * nBufferSize);
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), 2 * nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), (size_t)(2 * nBufferSize));
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)(2 * nBufferSize));
 	file.Close();
 
 	ASL_UTEST_ASSERT_TRUE(file.Open(szFilename, File::OF_In));
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), 0);
-	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), 2 * nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)0);
+	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), (size_t)(2 * nBufferSize));
 	ASL_UTEST_ASSERT_EQ(file.Read(acBuffer2, nBufferSize), nBufferSize);
 	ASL_UTEST_ASSERT_EQ(memcmp(acBuffer1, acBuffer2, nBufferSize), 0);
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)nBufferSize);
 	ASL_UTEST_ASSERT_EQ(file.Read(acBuffer2, nBufferSize), nBufferSize);
 	ASL_UTEST_ASSERT_EQ(memcmp(acBuffer1, acBuffer2, nBufferSize), 0);
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), 2 * nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)(2 * nBufferSize));
 	ASL_UTEST_ASSERT_EQ(file.Read(acBuffer2, nBufferSize), 0);
 	file.Close();
 
 	ASL_UTEST_ASSERT_TRUE(file.Open(szFilename, File::OF_In));
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), 0);
-	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), 2 * nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)0);
+	ASL_UTEST_ASSERT_EQ(file.GetFileSize(), (size_t)(2 * nBufferSize));
 	file.Seek(nBufferSize);
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)nBufferSize);
 	ASL_UTEST_ASSERT_EQ(file.Read(acBuffer2, nBufferSize), nBufferSize);
 	ASL_UTEST_ASSERT_EQ(memcmp(acBuffer1, acBuffer2, nBufferSize), 0);
-	ASL_UTEST_ASSERT_EQ(file.GetPos(), 2 * nBufferSize);
+	ASL_UTEST_ASSERT_EQ(file.GetPos(), (size_t)(2 * nBufferSize));
 	ASL_UTEST_ASSERT_EQ(file.Read(acBuffer2, nBufferSize), 0);
 	file.Close();
